backend/requirements.txt
Flask==3.1.3
flask-cors==6.0.2
groq==1.1.2
openai-whisper==20250625
PyJWT==2.12.1
python-dotenv==1.2.2
Werkzeug==3.1.6
gunicorn==23.0.0
